+++ src/main/java/dbservice/DAOs/WagonDAO.java
@@ -13,23 +13,23 @@ public class WagonDAO {
 		this.jdbcTemplate = jdbcTemplate;
 	}
 
-	public Long acquireID() {
+	public Integer acquireID() {
 		String sql = "SELECT TOP 1 wagon_id FROM Wagons ORDER BY wagon_id DESC";
-		return jdbcTemplate.queryForObject(sql, Long.class);
+		return jdbcTemplate.queryForObject(sql, Integer.class);
 	}
 
-	public Long addWagonGetID(Wagon wagon) {
+	public Integer addWagonGetID(Wagon wagon) {
 		String sql = "INSERT INTO Wagons (wagon_num, wagon_capacity, train_id) VALUES (?, ?, ?)";
 		jdbcTemplate.update(sql, wagon.getWagonNum(), wagon.getWagonCapacity(), wagon.getTrainID());
 		return acquireID();
 	}
 
-	public void deleteWagon(Long id) {
+	public void deleteWagon(Integer id) {
 		String sql = "DELETE FROM Wagons WHERE wagon_id = ?";
 		jdbcTemplate.update(sql, id);
 	}
 
-	public void changeWagonNum(Long id, int newWagonNum) {
+	public void changeWagonNum(Integer id, int newWagonNum) {
 		String sql = "UPDATE Wagons SET wagon_num = ? WHERE wagon_id = ?";
 		jdbcTemplate.update(sql, newWagonNum, id);
 	}
@@ -39,12 +39,12 @@ public class WagonDAO {
 		return jdbcTemplate.queryForObject(sql, int.class, id);
 	}
 
-	public Long getTrainID(Long id) {
+	public Integer getTrainID(Integer id) {
 		String sql = "SELECT train_id FROM Wagons WHERE wagon_id = ?";
-		return jdbcTemplate.queryForObject(sql, Long.class, id);
+		return jdbcTemplate.queryForObject(sql, Integer.class, id);
 	}
 
-	public void changeTrain(Long id, Long newTrainID) {
+	public void changeTrain(Integer id, Integer newTrainID) {
 		String sql = "UPDATE Wagons SET train_id = ? WHERE wagon_id = ?";
 		jdbcTemplate.update(sql, newTrainID, id);
 	}
