+++ src/main/java/dbservice/models/Ticket.java
@@ -4,11 +4,11 @@ import jakarta.persistence.Entity;
 
 @Entity
 public class Ticket {
-	private Long ticket_id;
-	private Long traveler_id;
-	private Long first_stop;
-	private Long last_stop;
-	private Long train_id;
+	private Integer ticket_id;
+	private Integer traveler_id;
+	private Integer first_stop;
+	private Integer last_stop;
+	private Integer train_id;
 	private int wagon_num;
 	private int seat_num;
 	private boolean is_expired;
@@ -16,7 +16,7 @@ public class Ticket {
 	public Ticket() {
 	}
 
-	public Ticket(Long traveler_id, Long first_stop, Long last_stop, Long train_id, int wagon_num, int seat_num) {
+	public Ticket(Integer traveler_id, Integer first_stop, Integer last_stop, Integer train_id, int wagon_num, int seat_num) {
 		this.traveler_id = traveler_id;
 		this.first_stop = first_stop;
 		this.last_stop = last_stop;
@@ -27,43 +27,43 @@ public class Ticket {
 	}
 
 	// getters and setters
-	public Long getTicketID() {
+	public Integer getTicketID() {
 		return ticket_id;
 	}
 
-	public void setTicketID(Long ticket_id) {
+	public void setTicketID(Integer ticket_id) {
 		this.ticket_id = ticket_id;
 	}
 
-	public Long getTravelerID() {
+	public Integer getTravelerID() {
 		return traveler_id;
 	}
 
-	public void setTravelerID(Long traveler_id) {
+	public void setTravelerID(Integer traveler_id) {
 		this.traveler_id = traveler_id;
 	}
 
-	public Long getFirstStop() {
+	public Integer getFirstStop() {
 		return first_stop;
 	}
 
-	public void setFirstStop(Long first_stop) {
+	public void setFirstStop(Integer first_stop) {
 		this.first_stop = first_stop;
 	}
 
-	public Long getLastStop() {
+	public Integer getLastStop() {
 		return last_stop;
 	}
 
-	public void setLastStop(Long last_stop) {
+	public void setLastStop(Integer last_stop) {
 		this.last_stop = last_stop;
 	}
 
-	public Long getTrainID() {
+	public Integer getTrainID() {
 		return train_id;
 	}
 
-	public void setTrainID(Long train_id) {
+	public void setTrainID(Integer train_id) {
 		this.train_id = train_id;
 	}
